Try fix

+++ install.sh
@@ -80,20 +80,21 @@ printc "  Check if node and npm is installed ...\n" "i"
 node --version > /dev/null 2>&1
 if [ ! $(echo $?) -eq 0 ]; then
 	curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.7/install.sh | bash
-	echo 'export NVM_DIR="$USER_HOME/.nvm"' >> $USER_HOME/.bashrc
-	echo '[ -s "$NVM_DIR/nvm.sh"  ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm' >> $USER_HOME/.bashrc
-	echo '[ -s "$NVM_DIR/bash_completion"  ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion' >> $USER_HOME/.bashrc
-	source $USER_HOME/.bashrc
+	export NVIM_DIR="$USER_HOME/.nvm"
+	echo "export NVM_DIR=$NVIM_DIR" >> $USER_HOME/.bashrc
+	echo "[ -s \"$NVM_DIR/nvm.sh\"  ] && \\. \"$NVM_DIR/nvm.sh\"  # This loads nvm" >> $USER_HOME/.bashrc
+	echo "[ -s \"$NVM_DIR/bash_completion\"  ] && \\. \"$NVM_DIR/bash_completion\"  # This loads nvm bash_completion" >> $USER_HOME/.bashrc
 	nvm install 14.18.1
 else
 	printc "  node and npm is installed. \n" "i"
 fi
 
-# .config/neovim
-printc "Install configurations for nvim and tmux\n" "s"
-git clone https://gitlab.com/lpg2709/dotfiles.git "/tmp/dotfiles"
-/tmp/dotfiles/install.sh --nvim --tmux --vim
-rm -rf "/tmp/dotfiles"
+if [ ! -d "$USER_HOME/.config/nvim" ]; then
+	printc "Install configurations for nvim and tmux\n" "s"
+	git clone https://gitlab.com/lpg2709/dotfiles.git "/tmp/dotfiles"
+	/tmp/dotfiles/install.sh --nvim --tmux --vim
+	rm -rf "/tmp/dotfiles"
+fi
 
 printc "Cleaning files ...\n" "s"
